Minimal virtio-gpu compilation fixes

+++ drivers/gpu/drm/virtio/virtgpu_debugfs.c
@@ -59,8 +59,8 @@ virtio_gpu_debugfs_irq_info(struct seq_file *m, void *data)
 	struct virtio_gpu_device *vgdev = node->minor->dev->dev_private;
 
 	seq_printf(m, "fence %llu %lld\n",
-		   (u64)atomic64_read(&vgdev->fence_drv.last_seq),
-		   vgdev->fence_drv.sync_seq);
+		   (unsigned long long)atomic64_read(&vgdev->fence_drv.last_seq),
+                   (long long)vgdev->fence_drv.sync_seq);
 	return 0;
 }
